requirements.txt
deepagents
tavily-python
langchain
langchain-community
langchain-experimental
docker
fastapi
uvicorn[standard]
sqlalchemy
psycopg[binary]
pydantic
pyyaml
claude-agent-sdk
PyJWT
email-validator
python-multipart
python-docx
openpyxl
python-pptx
pdfminer.six
Pillow
hindsight-client
